requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
httpx==0.24.1
pydantic==1.10.12
python-dotenv==1.0.0

# optional provider SDKs (install if you plan to use them)
openai>=1.0.0
anthropic>=0.33.0
google-generativeai>=0.2.0
